+++ rtl/sys_gen/tst_sram/nexys4d/sys_conf.vhd
@@ -1,0 +1,59 @@
+-- $Id: sys_conf.vhd 1099 2018-12-31 09:07:36Z mueller $
+--
+-- Copyright 2018- by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+--
+-- This program is free software; you may redistribute and/or modify it under
+-- the terms of the GNU General Public License as published by the Free
+-- Software Foundation, either version 3, or (at your option) any later version.
+--
+-- This program is distributed in the hope that it will be useful, but
+-- WITHOUT ANY WARRANTY, without even the implied warranty of MERCHANTABILITY
+-- or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+-- for complete details.
+--
+------------------------------------------------------------------------------
+-- Package Name:   sys_conf
+-- Description:    Definitions for sys_tst_sram_n4d (for synthesis)
+--
+-- Dependencies:   -
+-- Tool versions:  viv 2017.2; ghdl 0.34
+-- Revision History: 
+-- Date         Rev Version  Comment
+-- 2018-12-30  1099   1.0    Initial version 
+------------------------------------------------------------------------------
+
+library ieee;
+use ieee.std_logic_1164.all;
+
+use work.slvtypes.all;
+
+package sys_conf is
+
+  constant sys_conf_clksys_vcodivide   : positive :=   1;
+  constant sys_conf_clksys_vcomultiply : positive :=   1;   -- vco  --- MHz
+  constant sys_conf_clksys_outdivide   : positive :=   1;   -- sys  100 MHz
+  constant sys_conf_clksys_gentype     : string   := "MMCM";
+  -- dual clock design, clkser = 120 MHz
+  constant sys_conf_clkser_vcodivide   : positive :=   1;
+  constant sys_conf_clkser_vcomultiply : positive :=  12;   -- vco 1200 MHz
+  constant sys_conf_clkser_outdivide   : positive :=  10;   -- sys  120 MHz
+  constant sys_conf_clkser_gentype     : string   := "PLL";
+  
+  constant sys_conf_ser2rri_defbaud : integer := 115200;   -- default 115k baud
+
+  -- derived constants
+  
+  constant sys_conf_clksys : integer :=
+    ((100000000/sys_conf_clksys_vcodivide)*sys_conf_clksys_vcomultiply) /
+    sys_conf_clksys_outdivide;
+  constant sys_conf_clksys_mhz : integer := sys_conf_clksys/1000000;
+
+  constant sys_conf_clkser : integer :=
+     ((100000000/sys_conf_clkser_vcodivide)*sys_conf_clkser_vcomultiply) /
+    sys_conf_clkser_outdivide;
+  constant sys_conf_clkser_mhz : integer := sys_conf_clkser/1000000;
+
+  constant sys_conf_ser2rri_cdinit : integer :=
+    (sys_conf_clkser/sys_conf_ser2rri_defbaud)-1;
+
+end package sys_conf;